TFTGitBrokenLastUsefullPush
Due to a mishandling in github desktop stash feature, the proyect versioning didnt match the latest versions of the code. This is the last usefull commit done in the renamed repository "TFTGitBroken", corresponding with the push "Not rebootable" (commit a2504e74903c3a1cfd1f93b04a24ce9f527ca4a5) of the mentioned repository.